Move Spec to spec module
