+++ cubed/runtime/executors/lithops.py
@@ -20,7 +20,6 @@ from lithops.executors import FunctionExecutor
 from lithops.wait import ALWAYS, ANY_COMPLETED
 from networkx import MultiDiGraph
 
-from cubed.core.array import Spec
 from cubed.runtime.backup import should_launch_backup
 from cubed.runtime.executors.lithops_retries import (
     RetryingFunctionExecutor,
@@ -29,6 +28,7 @@ from cubed.runtime.executors.lithops_retries import (
 from cubed.runtime.pipeline import visit_node_generations, visit_nodes
 from cubed.runtime.types import Callback, DagExecutor
 from cubed.runtime.utils import handle_callbacks
+from cubed.spec import Spec
 
 logger = logging.getLogger(__name__)
