+++ cubed/runtime/executors/modal_async.py
@@ -9,7 +9,6 @@ from modal.functions import Function
 from networkx import MultiDiGraph
 from tenacity import retry, retry_if_exception_type, stop_after_attempt
 
-from cubed.core.array import Spec
 from cubed.runtime.executors.asyncio import async_map_unordered
 from cubed.runtime.executors.modal import (
     Container,
@@ -20,6 +19,7 @@ from cubed.runtime.executors.modal import (
 from cubed.runtime.pipeline import visit_node_generations, visit_nodes
 from cubed.runtime.types import Callback, DagExecutor
 from cubed.runtime.utils import handle_callbacks
+from cubed.spec import Spec
 
 
 # We need map_unordered for the use_backups implementation
